compile quaint to wasm
