+++ quaint/Cargo.toml
@@ -23,18 +23,26 @@ resolver = "2"
 features = ["docs", "all"]
 
 [features]
-default = []
+default = ["slim"]
 docs = []
 # Expose the underlying database drivers when a connector is enabled. This is a
 # way to access database-specific methods when you need extra control.
 expose-drivers = []
 
-all = [
+all = ["slim", "connector", "pooled"]
+
+connector = [
+  "postgresql-connector",
+  "mysql-connector",
+  "mssql-connector",
+  "sqlite-connector",
+]
+
+slim = [
   "chrono",
   "json",
   "mssql",
   "mysql",
-  "pooled",
   "postgresql",
   "serde-support",
   "sqlite",
@@ -47,7 +55,8 @@ vendored-openssl = [
   "mysql_async/vendored-openssl",
 ]
 
-postgresql = [
+postgresql-connector = [
+  "postgresql",
   "native-tls",
   "tokio-postgres",
   "postgres-types",
@@ -58,13 +67,25 @@ postgresql = [
   "lru-cache",
   "byteorder",
 ]
+postgresql = []
 
 json = ["serde_json", "base64"]
-mssql = ["tiberius", "uuid", "chrono", "tokio-util", "tokio/time", "tokio/net", "either"]
-mysql = ["mysql_async", "tokio/time", "lru-cache"]
+mssql-connector = [
+  "mssql",
+  "tiberius",
+  "uuid",
+  "chrono",
+  "tokio-util",
+  "tokio/time",
+  "tokio/net",
+]
+mssql = []
+mysql-connector = ["mysql", "mysql_async", "tokio/time", "lru-cache"]
+mysql = ["chrono/std"]
 pooled = ["mobc"]
 serde-support = ["serde", "chrono/serde"]
-sqlite = ["rusqlite", "tokio/sync"]
+sqlite-connector = ["sqlite", "rusqlite", "tokio/sync"]
+sqlite = []
 bigdecimal = ["bigdecimal_"]
 fmt-sql = ["sqlformat"]
 
@@ -81,11 +102,13 @@ futures = "0.3"
 url = "2.1"
 hex = "0.4"
 
-either = { version = "1.6", optional = true }
+either = "1.6"
 base64 = { version = "0.12.3", optional = true }
 chrono = { version = "0.4", optional = true, default-features = false }
 lru-cache = { version = "0.1", optional = true }
-serde_json = { version = "1.0.48", optional = true, features = ["float_roundtrip"] }
+serde_json = { version = "1.0.48", optional = true, features = [
+  "float_roundtrip",
+] }
 native-tls = { version = "0.2", optional = true }
 bit-vec = { version = "0.6.1", optional = true }
 bytes = { version = "1.0", optional = true }
@@ -103,6 +126,10 @@ test-macros = { path = "test-macros" }
 test-setup = { path = "test-setup" }
 uuid = { version = "1", features = ["v4"] }
 tokio = { version = "1.0", features = ["rt-multi-thread", "macros", "time"] }
+
+[target.'cfg(target_arch = "wasm32")'.dependencies.getrandom]
+version = "0.2"
+features = ["js"]
 
 [dependencies.byteorder]
 default-features = false
@@ -128,7 +155,14 @@ features = ["sql-browser-tokio", "chrono", "bigdecimal"]
 version = "0.11.2"
 optional = true
 default-features = false
-features = ["sql-browser-tokio", "vendored-openssl", "chrono", "bigdecimal", "tds73", "winauth"]
+features = [
+  "sql-browser-tokio",
+  "vendored-openssl",
+  "chrono",
+  "bigdecimal",
+  "tds73",
+  "winauth",
+]
 
 [dependencies.bigdecimal_]
 version = "0.3"
@@ -140,13 +174,23 @@ version = "1"
 optional = true
 
 [dependencies.tokio-postgres]
-features = ["with-uuid-1", "with-chrono-0_4", "with-serde_json-1", "with-bit-vec-0_6"]
+features = [
+  "with-uuid-1",
+  "with-chrono-0_4",
+  "with-serde_json-1",
+  "with-bit-vec-0_6",
+]
 git = "https://github.com/prisma/rust-postgres"
 branch = "pgbouncer-mode"
 optional = true
 
 [dependencies.postgres-types]
-features = ["with-uuid-1", "with-chrono-0_4", "with-serde_json-1", "with-bit-vec-0_6"]
+features = [
+  "with-uuid-1",
+  "with-chrono-0_4",
+  "with-serde_json-1",
+  "with-bit-vec-0_6",
+]
 git = "https://github.com/prisma/rust-postgres"
 branch = "pgbouncer-mode"
 optional = true
