+++ quaint/src/ast/update.rs
@@ -150,10 +150,7 @@ impl<'a> Update<'a> {
     /// # }
     /// ```
     #[cfg(any(feature = "postgresql", feature = "sqlite"))]
-    #[cfg_attr(
-        feature = "docs",
-        doc(cfg(any(feature = "postgresql", feature = "sqlite")))
-    )]
+    #[cfg_attr(feature = "docs", doc(cfg(any(feature = "postgresql", feature = "sqlite"))))]
     pub fn returning<K, I>(mut self, columns: I) -> Self
     where
         K: Into<Column<'a>>,
