+++ quaint/src/connector/mssql.rs
@@ -1,21 +1,18 @@
 mod conversion;
 mod error;
 
+use super::mssql_common::*;
 use super::{IsolationLevel, TransactionOptions};
 use crate::{
     ast::{Query, Value},
     connector::{metrics, queryable::*, ResultSet, Transaction},
-    error::{Error, ErrorKind},
     visitor::{self, Visitor},
 };
 use async_trait::async_trait;
-use connection_string::JdbcString;
 use futures::lock::Mutex;
 use std::{
     convert::TryFrom,
-    fmt,
     future::Future,
-    str::FromStr,
     sync::atomic::{AtomicBool, Ordering},
     time::Duration,
 };
@@ -26,71 +23,6 @@ use tokio_util::compat::{Compat, TokioAsyncWriteCompatExt};
 /// The underlying SQL Server driver. Only available with the `expose-drivers` Cargo feature.
 #[cfg(feature = "expose-drivers")]
 pub use tiberius;
-
-/// Wraps a connection url and exposes the parsing logic used by Quaint,
-/// including default values.
-#[derive(Debug, Clone)]
-#[cfg_attr(feature = "docs", doc(cfg(feature = "mssql")))]
-pub struct MssqlUrl {
-    connection_string: String,
-    query_params: MssqlQueryParams,
-}
-
-/// TLS mode when connecting to SQL Server.
-#[derive(Debug, Clone, Copy)]
-#[cfg_attr(feature = "docs", doc(cfg(feature = "mssql")))]
-pub enum EncryptMode {
-    /// All traffic is encrypted.
-    On,
-    /// Only the login credentials are encrypted.
-    Off,
-    /// Nothing is encrypted.
-    DangerPlainText,
-}
-
-impl fmt::Display for EncryptMode {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::On => write!(f, "true"),
-            Self::Off => write!(f, "false"),
-            Self::DangerPlainText => write!(f, "DANGER_PLAINTEXT"),
-        }
-    }
-}
-
-impl FromStr for EncryptMode {
-    type Err = Error;
-
-    fn from_str(s: &str) -> crate::Result<Self> {
-        let mode = match s.parse::<bool>() {
-            Ok(true) => Self::On,
-            _ if s == "DANGER_PLAINTEXT" => Self::DangerPlainText,
-            _ => Self::Off,
-        };
-
-        Ok(mode)
-    }
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct MssqlQueryParams {
-    encrypt: EncryptMode,
-    port: Option<u16>,
-    host: Option<String>,
-    user: Option<String>,
-    password: Option<String>,
-    database: String,
-    schema: String,
-    trust_server_certificate: bool,
-    trust_server_certificate_ca: Option<String>,
-    connection_limit: Option<usize>,
-    socket_timeout: Option<Duration>,
-    connect_timeout: Option<Duration>,
-    pool_timeout: Option<Duration>,
-    transaction_isolation_level: Option<IsolationLevel>,
-    max_connection_lifetime: Option<Duration>,
-    max_idle_connection_lifetime: Option<Duration>,
-}
 
 static SQL_SERVER_DEFAULT_ISOLATION: IsolationLevel = IsolationLevel::ReadCommitted;
 
@@ -108,158 +40,6 @@ impl TransactionCapable for Mssql {
         let opts = TransactionOptions::new(isolation, self.requires_isolation_first());
 
         Transaction::new(self, self.begin_statement(), opts).await
-    }
-}
-
-impl MssqlUrl {
-    /// Maximum number of connections the pool can have (if used together with
-    /// pooled Quaint).
-    pub fn connection_limit(&self) -> Option<usize> {
-        self.query_params.connection_limit()
-    }
-
-    /// A duration how long one query can take.
-    pub fn socket_timeout(&self) -> Option<Duration> {
-        self.query_params.socket_timeout()
-    }
-
-    /// A duration how long we can try to connect to the database.
-    pub fn connect_timeout(&self) -> Option<Duration> {
-        self.query_params.connect_timeout()
-    }
-
-    /// A pool check_out timeout.
-    pub fn pool_timeout(&self) -> Option<Duration> {
-        self.query_params.pool_timeout()
-    }
-
-    /// The isolation level of a transaction.
-    fn transaction_isolation_level(&self) -> Option<IsolationLevel> {
-        self.query_params.transaction_isolation_level
-    }
-
-    /// Name of the database.
-    pub fn dbname(&self) -> &str {
-        self.query_params.database()
-    }
-
-    /// The prefix which to use when querying database.
-    pub fn schema(&self) -> &str {
-        self.query_params.schema()
-    }
-
-    /// Database hostname.
-    pub fn host(&self) -> &str {
-        self.query_params.host()
-    }
-
-    /// The username to use when connecting to the database.
-    pub fn username(&self) -> Option<&str> {
-        self.query_params.user()
-    }
-
-    /// The password to use when connecting to the database.
-    pub fn password(&self) -> Option<&str> {
-        self.query_params.password()
-    }
-
-    /// The TLS mode to use when connecting to the database.
-    pub fn encrypt(&self) -> EncryptMode {
-        self.query_params.encrypt()
-    }
-
-    /// If true, we allow invalid certificates (self-signed, or otherwise
-    /// dangerous) when connecting. Should be true only for development and
-    /// testing.
-    pub fn trust_server_certificate(&self) -> bool {
-        self.query_params.trust_server_certificate()
-    }
-
-    /// Path to a custom server certificate file.
-    pub fn trust_server_certificate_ca(&self) -> Option<&str> {
-        self.query_params.trust_server_certificate_ca()
-    }
-
-    /// Database port.
-    pub fn port(&self) -> u16 {
-        self.query_params.port()
-    }
-
-    /// The JDBC connection string
-    pub fn connection_string(&self) -> &str {
-        &self.connection_string
-    }
-
-    /// The maximum connection lifetime
-    pub fn max_connection_lifetime(&self) -> Option<Duration> {
-        self.query_params.max_connection_lifetime()
-    }
-
-    /// The maximum idle connection lifetime
-    pub fn max_idle_connection_lifetime(&self) -> Option<Duration> {
-        self.query_params.max_idle_connection_lifetime()
-    }
-}
-
-impl MssqlQueryParams {
-    fn port(&self) -> u16 {
-        self.port.unwrap_or(1433)
-    }
-
-    fn host(&self) -> &str {
-        self.host.as_deref().unwrap_or("localhost")
-    }
-
-    fn user(&self) -> Option<&str> {
-        self.user.as_deref()
-    }
-
-    fn password(&self) -> Option<&str> {
-        self.password.as_deref()
-    }
-
-    fn encrypt(&self) -> EncryptMode {
-        self.encrypt
-    }
-
-    fn trust_server_certificate(&self) -> bool {
-        self.trust_server_certificate
-    }
-
-    fn trust_server_certificate_ca(&self) -> Option<&str> {
-        self.trust_server_certificate_ca.as_deref()
-    }
-
-    fn database(&self) -> &str {
-        &self.database
-    }
-
-    fn schema(&self) -> &str {
-        &self.schema
-    }
-
-    fn socket_timeout(&self) -> Option<Duration> {
-        self.socket_timeout
-    }
-
-    fn connect_timeout(&self) -> Option<Duration> {
-        self.connect_timeout
-    }
-
-    fn connection_limit(&self) -> Option<usize> {
-        self.connection_limit
-    }
-
-    fn pool_timeout(&self) -> Option<Duration> {
-        self.pool_timeout
-    }
-
-    fn max_connection_lifetime(&self) -> Option<Duration> {
-        self.max_connection_lifetime
-    }
-
-    fn max_idle_connection_lifetime(&self) -> Option<Duration> {
-        self.max_idle_connection_lifetime
     }
 }
 
@@ -447,150 +227,6 @@ impl Queryable for Mssql {
 
     fn requires_isolation_first(&self) -> bool {
         true
-    }
-}
-
-impl MssqlUrl {
-    pub fn new(jdbc_connection_string: &str) -> crate::Result<Self> {
-        let query_params = Self::parse_query_params(jdbc_connection_string)?;
-        let connection_string = Self::with_jdbc_prefix(jdbc_connection_string);
-
-        Ok(Self {
-            connection_string,
-            query_params,
-        })
-    }
-
-    fn with_jdbc_prefix(input: &str) -> String {
-        if input.starts_with("jdbc:sqlserver") {
-            input.into()
-        } else {
-            format!("jdbc:{input}")
-        }
-    }
-
-    fn parse_query_params(input: &str) -> crate::Result<MssqlQueryParams> {
-        let mut conn = JdbcString::from_str(&Self::with_jdbc_prefix(input))?;
-
-        let host = conn.server_name().map(|server_name| match conn.instance_name() {
-            Some(instance_name) => format!(r#"{server_name}\{instance_name}"#),
-            None => server_name.to_string(),
-        });
-
-        let port = conn.port();
-        let props = conn.properties_mut();
-        let user = props.remove("user");
-        let password = props.remove("password");
-        let database = props.remove("database").unwrap_or_else(|| String::from("master"));
-        let schema = props.remove("schema").unwrap_or_else(|| String::from("dbo"));
-
-        let connection_limit = props
-            .remove("connectionlimit")
-            .or_else(|| props.remove("connection_limit"))
-            .map(|param| param.parse())
-            .transpose()?;
-
-        let transaction_isolation_level = props
-            .remove("isolationlevel")
-            .or_else(|| props.remove("isolation_level"))
-            .map(|level| {
-                IsolationLevel::from_str(&level).map_err(|_| {
-                    let kind = ErrorKind::database_url_is_invalid(format!("Invalid isolation level `{level}`"));
-                    Error::builder(kind).build()
-                })
-            })
-            .transpose()?;
-
-        let mut connect_timeout = props
-            .remove("logintimeout")
-            .or_else(|| props.remove("login_timeout"))
-            .or_else(|| props.remove("connecttimeout"))
-            .or_else(|| props.remove("connect_timeout"))
-            .or_else(|| props.remove("connectiontimeout"))
-            .or_else(|| props.remove("connection_timeout"))
-            .map(|param| param.parse().map(Duration::from_secs))
-            .transpose()?;
-
-        match connect_timeout {
-            None => connect_timeout = Some(Duration::from_secs(5)),
-            Some(dur) if dur.as_secs() == 0 => connect_timeout = None,
-            _ => (),
-        }
-
-        let mut pool_timeout = props
-            .remove("pooltimeout")
-            .or_else(|| props.remove("pool_timeout"))
-            .map(|param| param.parse().map(Duration::from_secs))
-            .transpose()?;
-
-        match pool_timeout {
-            None => pool_timeout = Some(Duration::from_secs(10)),
-            Some(dur) if dur.as_secs() == 0 => pool_timeout = None,
-            _ => (),
-        }
-
-        let socket_timeout = props
-            .remove("sockettimeout")
-            .or_else(|| props.remove("socket_timeout"))
-            .map(|param| param.parse().map(Duration::from_secs))
-            .transpose()?;
-
-        let encrypt = props
-            .remove("encrypt")
-            .map(|param| EncryptMode::from_str(&param))
-            .transpose()?
-            .unwrap_or(EncryptMode::On);
-
-        let trust_server_certificate = props
-            .remove("trustservercertificate")
-            .or_else(|| props.remove("trust_server_certificate"))
-            .map(|param| param.parse())
-            .transpose()?
-            .unwrap_or(false);
-
-        let trust_server_certificate_ca: Option<String> = props
-            .remove("trustservercertificateca")
-            .or_else(|| props.remove("trust_server_certificate_ca"));
-
-        let mut max_connection_lifetime = props
-            .remove("max_connection_lifetime")
-            .map(|param| param.parse().map(Duration::from_secs))
-            .transpose()?;
-
-        match max_connection_lifetime {
-            Some(dur) if dur.as_secs() == 0 => max_connection_lifetime = None,
-            _ => (),
-        }
-
-        let mut max_idle_connection_lifetime = props
-            .remove("max_idle_connection_lifetime")
-            .map(|param| param.parse().map(Duration::from_secs))
-            .transpose()?;
-
-        match max_idle_connection_lifetime {
-            None => max_idle_connection_lifetime = Some(Duration::from_secs(300)),
-            Some(dur) if dur.as_secs() == 0 => max_idle_connection_lifetime = None,
-            _ => (),
-        }
-
-        Ok(MssqlQueryParams {
-            encrypt,
-            port,
-            host,
-            user,
-            password,
-            database,
-            schema,
-            trust_server_certificate,
-            trust_server_certificate_ca,
-            connection_limit,
-            socket_timeout,
-            connect_timeout,
-            pool_timeout,
-            transaction_isolation_level,
-            max_connection_lifetime,
-            max_idle_connection_lifetime,
-        })
     }
 }
